ABC.Interact.dat {dae}	R Documentation

Randomly generated set of values indexed by three factors

Description

This data set has randomly generated values of the response variable MOE (Measure Of Effectiveness) which is indexed by the two-level factors A, B and C.

Usage

data(ABC.Interact.dat)

Format

A data.frame containing 8 observations of 4 variables.

Source

Generated by Chris Brien

[Package dae version 3.2.32 Index]