BIBDWheat.dat {dae}	R Documentation

Data for a balanced incomplete block experiment

Description

The data set comes from Joshi (1987) and is the data from an experiment to investigate six varieties of wheat that employs a balanced incomplete block design (BIBD) with ten blocks, each consisting of three plots. For more details see the vignette accessed via vignette("DesignNotes", package="dae").

Usage

data(BIBDWheat.dat)

Format

A data.frame containing 30 observations of 4 variables.

Source

Joshi, D. D. (1987) Linear Estimation and Design of Experiments. Wiley Eastern, New Delhi.

[Package dae version 3.2.32 Index]