Cabinet1.des {dae}	R Documentation

A design for one of the growth cabinets in an experiment with 50 lines and 4 harvests

Description

The systematic design for a lattice square for 49 treatments consisting of four 7 x 7 squares. For more details see the vignette daeDesignNotes.pdf.

Usage

data(Cabinet1.des)

Format

A data.frame containing 160 observations of 15 variables.

[Package dae version 3.2.32 Index]