Exp249.munit.des {dae}	R Documentation

Systematic, main-unit design for an experiment to be run in a greenhouse

Description

In this main-unit design, there are 24 lanes by 11 Positions, the lanes being blocked into 6 Zones of 4 lanes. The design for the main units is for assigning 75 wheat lines, of which 73 are Nested Association Mapping (NAM) wheat lines and the other two are two check lines, Scout and Gladius. A row and column design was generated with DiGGer (Coombes, 2009). For more details see the vignette accessed via vignette("DesignNotes", package="dae").

Usage

data(Exp249.munit.des)

Format

A data.frame containing 264 observations of 3 variables.

Source

Coombes, N. E. (2009) Digger: design search tool in R. URL: http://nswdpibiom.org/austatgen/software/, (accessed June 3, 2015).

[Package dae version 3.2.32 Index]