Fac4Proc.dat {dae}	R Documentation

Data for a 2^4 factorial experiment

Description

The data set come from an unreplicated 2^4 factorial experiment to investigate a chemical process. The response variable is the Conversion percentage (Conv) and this is indexed by the 4 two-level factors Catal, Temp, Press and Conc, with levels “-” and “+”. The data is aranged in Yates order. Also included is the 16-level factor Runs which gives the order in which the combinations of the two-level factors were run.

Usage

data(Fac4Proc.dat)

Format

A data.frame containing 16 observations of 6 variables.

Source

Table 10.6 of Box, Hunter and Hunter (1978) Statistics for Experimenters. New York, Wiley.

[Package dae version 3.2.32 Index]