LatticeSquare_t49.des {dae}	R Documentation

A Lattice square design for 49 treatments

Description

The systematic design for a lattice square for 49 treatments consisting of four 7 x 7 squares. For more details see the vignette accessed via vignette("DesignNotes", package="dae").

Usage

data(LatticeSquare_t49.des)

Format

A data.frame containing 196 observations of 4 variables.

Source

Cochran and Cox (1957) Experimental Designs. 2nd edn Wiley, New York.

[Package dae version 3.2.32 Index]