McIntyreTMV.dat {dae}	R Documentation

The design and data from McIntyre's (1955) two-phase experiment

Description

McIntyre (1955) reports an investigation of the effect of four light intensities on the synthesis of tobacco mosaic virus in leaves of tobacco Nicotiana tabacum var. Hickory Pryor. It is a two-phase experiment: the first phase is a treatment phase, in which the four light treatments are randomized to the tobacco leaves, and the second phase is an assay phase, in which the tobacco leaves are randomized to the half-leaves of assay plants. For more details see the vignette accessed via vignette("DesignNotes", package="dae").

Usage

data(McIntyreTMV.dat)

Format

A data.frame containing 196 observations of 4 variables.

Source

McIntyre, G. A. (1955) Design and Analysis of Two Phase Experiments. Biometrics, 11, 324–334.

[Package dae version 3.2.32 Index]