Zncsspline {dae}	R Documentation

Calculates the design matrix for fitting the random component of a natural cubic smoothing spline

Description

Calculates the design matrix, Z_S, of the random effects for a natural cubic smoothing spline as described by Verbyla et al., (1999). An initial design matrix, \bold{Delta} \bold{Delta}^{-1} \bold{Delta}, based on the knot points is computed. It can then be post multiplied by a power of the tri-diagonal matrix G_S, G_S being proportional to the assumed variance matrix of the random spline effects. If the power is set to 0.5, then the random spline effects based on the resulting design matrix Z_S are now independent with variance sigma_s^2. The variance component that estimates sigma_s^2 will then be a variance ratio and the smoothing parameter is the inverse of the ratio of this variance component to the residual variance.

Usage

Zncsspline(knot.points, Gpower = 0, print = FALSE)

Arguments

knot.points
A numeric giving the values of the knot points to be used in fitting the spline. These must be orderd in increasing order.

Gpower
A numeric giving the power to take of the tri-diagonal matrix G_s; the initial design matrix is postmultiplied by the G_s raised to Gpower to produce the design matrix Z_S that is returned.

print
A logical indicating whether to print the Delta and G_s matrices.

Value

A matrix that is the design matrix Z_S.

Author(s)

Chris Brien

References

Verbyla, A. P., Cullis, B. R., Kenward, M. G., and Welham, S. J. (1999). The analysis of designed experiments and longitudinal data by using smoothing splines (with discussion). Journal of the Royal Statistical Society, Series C (Applied Statistics), 48, 269-311.

See Also

mat.ncssvar.

Examples

Z <- Zncsspline(knot.points = 1:10, Gpower = 0.5)

[Package dae version 3.2.32 Index]