dae-deprecated {dae}	R Documentation

Deprecated Functions in Package dae

Description

These functions have been renamed and deprecated in dae.

Usage

Ameasures(...)
blockboundary.plot(...)
design.plot(...)
proj2.decomp(...)
proj2.ops(...)
projs.canon(...)
projs.structure(...)

Arguments

...
absorbs arguments passed from the old functions of the style foo.bar().

Author(s)

Chris Brien

[Package dae version 3.2.32 Index]