dae-package {dae}	R Documentation

Functions Useful in the Design and ANOVA of Experiments

Description

The content falls into the following groupings: (i) Data, (ii) Factor manipulation functions, (iii) Design functions, (iv) ANOVA functions, (v) Matrix functions, (vi) Projector and canonical efficiency functions, and (vii) Miscellaneous functions. There is a vignette describing how to use the design functions for randomizing and assessing designs available as a vignette called 'DesignNotes'. The ANOVA functions facilitate the extraction of information when the 'Error' function has been used in the call to 'aov'. The package 'dae' can also be installed from <http://chris.brien.name/rpackages/>.

Version: 3.2.32

Date: 2025-08-20

Index

(i) Data

ABC.Interact.dat 	 Randomly generated set of values indexed by

three factors

BIBDWheat.dat 	 Data for a balanced incomplete block experiment

Casuarina.dat 	 Data for an experiment with rows and columns from

Williams (2002)

Exp249.munit.des 	 Systematic, main-plot design for an experiment to be

run in a greenhouse

Fac4Proc.dat 	 Data for a 2^4 factorial experiment

LatticeSquare_t49.des 	 A Lattice square design for 49 treatments

McIntyreTMV.dat 	 The design and data from McIntyre (1955) two-phase

experiment

Oats.dat 	 Data for an experiment to investigate nitrogen response

of 3 oats varieties

Sensory3Phase.dat 	 Data for the three-phahse sensory evaluation

experiment in Brien, C.J. and Payne, R.W. (1999)

Sensory3PhaseShort.dat 	 Data for the three-phase sensory evaluation

experiment in Brien, C.J. and Payne, R.W. (1999),

but with short factor names

SPLGrass.dat 	 Data for an experiment to investigate the

effects of grazing patterns on pasture

composition

(ii) Factor manipulation functions

Forms a new or revised factor:

fac.combine 	 Combines several factors into one

fac.nested 	 Creates a factor, the nested factor, whose values are

generated within those of a nesting factor

fac.recast 	 Recasts a factor by modifying the values in the factor vector

and/or the levels attribute, possibly combining

some levels into a single level.

fac.recode 	 Recodes factor 'levels' using possibly nonunique

values in a vector. (May be deprecated in future.)

fac.uselogical 	 Forms a two-level factor from a logical object

Forms multiple new factors:

fac.divide 	 Divides a factor into several separate factors

fac.gen 	 Generate all combinations of several factors and,

optionally, replicate them

fac.genfactors 	 Generate all combinations of the levels of the supplied

factors, without replication

fac.multinested 	 Creates several factors, one for each level of the nesting factor

and each of whose values are either generated within those of

a level of the nesting factor or using the values of the nested

factor within the levels of the nesting factor.

fac.split 	 Splits a factor whose levels consist of several delimited

strings into several factors.

fac.uncombine 	 Cleaves a single factor, each of whose levels has delimited

strings, into several factors using the separated strings.

Operates on factors:

as.numfac 	 Convert a factor to a numeric vector, possibly centering or

scaling the values

mpone 	 Converts the first two levels of a factor into

the numeric values -1 and +1

fac.match 	 Match, for each combination of a set of columns

in 'x', the row that has the same combination

in 'table'

(iii) Design functions

Designing experiments:

designLatinSqrSys 	 Generate a systematic plan for a Latin Square design.

designRandomize 	 Randomize allocated to recipient factors to produce

a layout for an experiment. It supersedes fac.layout.

no.reps 	 Computes the number of replicates for an experiment

detect.diff 	 Computes the detectable difference for an experiment

power.exp 	 Computes the power for an experiment

Plotting designs:

blockboundaryPlot 	 This function plots a block boundary on a plot

produced by 'designPlot'. It supersedes

blockboundary.plot.

designBlocksGGPlot 	 Adds block boundaries to a plot produced by designGGPlot.

designGGPlot 	 Plots labels on a two-way grid of coloured cells using ggplot2

to represent an experimental design.

designPlot 	 A graphical representation of an experimental design

using labels stored in a matrix.

It superseded design.plot.

designPlotlabels 	 Plots labels on a two-way grid using ggplot2.

Assessing designs:

designAmeasures 	 Calculates the A-optimality measures from the

variance matrix for predictions.

designAnatomy 	 Given the layout for a design, obtain its anatomy via

the canonical analysis of its projectors to show the

confounding and aliasing inherent in the design.

designTwophaseAnatomies 	 Given the layout for a design and three structure formulae,

obtain the anatomies for the (i) two-phase,

(ii) first-phase, (iii) cross-phase, treatments, and

(iv) combined-units designs.

marginality.pstructure 	 Extracts the marginality matrix from a

pstructure.object

marginality.pstructure 	 Extracts a list containing the marginality matrices from

a pcanon.object

print.aliasing 	 Prints an aliasing data.frame

summary.pcanon 	 Summarizes the anatomy of a design, being the

decomposition of the sample space based on its

canonical analysis.

(iv) ANOVA functions

fitted.aovlist 	 Extract the fitted values for a fitted model

from an aovlist object

fitted.errors 	 Extract the fitted values for a fitted model

interaction.ABC.plot 	 Plots an interaction plot for three factors

qqyeffects 	 Half or full normal plot of Yates effects

resid.errors 	 Extract the residuals for a fitted model

residuals.aovlist 	 Extract the residuals from an aovlist object

strength 	 Generate paper strength values

tukey.1df 	 Performs Tukey's

one-degree-of-freedom-test-for-nonadditivity

yates.effects 	 Extract Yates effects

(v) Matrix functions

Operates on matrices:

elements 	 Extract the elements of an array specified by

the subscripts

mat.dirprod 	 Forms the direct product of two matrices

mat.dirsum 	 Forms the direct sum of a list of matrices

mat.ginv 	 Computes the generalized inverse of a matrix

Zncsspline 	 Forms the design matrix for fitting the

random effects for a natural cubic smoothing

spline.

Compute variance matrices for

supplied variance component values:

mat.random 	 Calculates the variance matrix for the

random effects from a mixed model, based

on a formula or a supplied matrix

mat.Vpred 	 Forms the variance matrix of predictions

based on supplied matrices

mat.Vpredicts 	 Forms the variance matrix of predictions,

based on supplied matrices or formulae.

Forms matrices using factors

stored in a data.frame:

fac.ar1mat 	 Forms the ar1 correlation matrix for a

(generalized) factor

fac.sumop 	 Computes the summation matrix that produces

sums corresponding to a (generalized) factor

fac.vcmat 	 Forms the variance matrix for the variance

component of a (generalized) factor

Forms patterned matrices:

mat.I 	 Forms a unit matrix

mat.J 	 Forms a square matrix of ones

mat.ncssvar 	 Forms a variance matrix for random cubic

smoothing spline effects

Forms correlation matrices:

mat.cor 	 Forms a correlation matrix in which all

correlations have the same value

mat.corg 	 Forms a general correlation matrix in which

all correlations have different values

mat.ar1 	 Forms an ar1 correlation matrix

mat.ar2 	 Forms an ar2 correlation matrix

mat.ar3 	 Forms an ar3 correlation matrix

mat.arma 	 Forms an arma correlation matrix

mat.banded 	 Forms a banded matrix

mat.exp 	 Forms an exponential correlation matrix

mat.gau 	 Forms a gaussian correlation matrix

mat.ma1 	 Forms an ma1 correlation matrix

mat.ma2 	 Forms an ma2 correlation matrix

mat.sar 	 Forms an sar correlation matrix

mat.sar2 	 Forms an sar2 correlation matrix

(vi) Projector and canonical efficiency functions

Projector class:

projector 	 Create projectors

projector-class 	 Class projector

is.projector 	 Tests whether an object is a valid object of

class projector

print.projector 	 Print projectors

correct.degfree 	 Check the degrees of freedom in an object of

class projector

degfree 	 Degrees of freedom extraction and replacement

Accepts two or more formulae:

designAnatomy 	 An anatomy of a design, obtained from

a canonical analysis of the relationships

between sets of projectors.

summary.pcanon 	 Summarizes the anatomy of a design, being the

decomposition of the sample space based on its

canonical analysis

print.summary.pcanon 	 Prints the values in an 'summary.pcanon' object

efficiencies.pcanon 	 Extracts the canonical efficiency factors from a

list of class 'pcanon'

Accepts exactly two formulae:

projs.2canon 	 A canonical analysis of the relationships between

two sets of projectors

projs.combine.p2canon 	 Extract, from a p2canon object, the projectors

summary.p2canon 	 A summary of the results of an analysis of

the relationships between two sets of projectors

print.summary.p2canon 	 Prints the values in an 'summary.p2canon' object

that give the combined decomposition

efficiencies.p2canon 	 Extracts the canonical efficiency factors from

a list of class 'p2canon'

Accepts a single formula:

as.data.frame.pstructure 	 Coerces a pstructure.object to a data.frame

print.pstructure 	 Prints a pstructure.object

pstructure.formula 	 Takes a formula and constructs a pstructure.object

that includes the orthogonalized projectors for the

terms in a formula

porthogonalize.list 	 Takes a list of projectors and constructs

a pstructure.object that includes projectors,

each of which has been orthogonalized to all projectors

preceding it in the list.

Others:

decomp.relate 	 Examines the relationship between the

eigenvectors for two decompositions

efficiency.criteria 	 Computes efficiency criteria from a set of

efficiency factors

fac.meanop 	 Computes the projection matrix that produces means

proj2.eigen 	 Canonical efficiency factors and eigenvectors

in joint decomposition of two projectors

proj2.efficiency 	 Computes the canonical efficiency factors for

the joint decomposition of two projectors

proj2.combine 	 Compute the projection and Residual operators

for two, possibly nonorthogonal, projectors

show-methods 	 Methods for Function 'show' in Package dae

(vii) Miscellaneous functions

extab 	 Expands the values in table to a vector

get.daeRNGkind 	 Gets the value of daeRNGkind for the package dae from

the daeEnv environment.

get.daeTolerance 	 Gets the value of daeTolerance for the package dae.

harmonic.mean 	 Calcuates the harmonic mean.

is.allzero 	 Tests whether all elements are approximately zero

rep.data.frame 	 Replicate the rows of a data.frame by repeating

each row consecutively and/or repeating all rows

as a group.

rmvnorm 	 Generates a vector of random values from a

multivariate normal distribution

set.daeRNGkind 	 Sets the values of daeRNGkind for the package dae in

the daeEnv environment'

set.daeTolerance 	 Sets the value of daeTolerance for the package dae.

Author(s)

Chris Brien [aut, cre] (ORCID: <https://orcid.org/0000-0003-0581-1817>)

Maintainer: Chris Brien <chris.brien@adelaide.edu.au>

[Package dae version 3.2.32 Index]