designLatinSqrSys {dae}	R Documentation

Generate a systematic plan for a Latin Square design

Description

Generates a systematic plan for a Latin Square design using the method of cycling the integers 1 to the number of treatments. The start of the cycle for each row, or the first column, can be specified as a vector of integers.

Usage

designLatinSqrSys(order, start = NULL)

Arguments

order
The number of treatments.

start
A numeric containing order unique values between one and order. These are interpreted as the value for the fist column for each row. If NULL, 1:order is used.

Value

A numeric containing order x order integers between 1 and order such that, when the numeric is considered as a square matrix of size order, each integer occurs once and only once in each row and column of the matrix.

See Also

designRandomize, designPlot, designAnatomy in package dae.

Examples

   matrix(designLatinSqrSys(5, start = c(seq(1, 5, 2), seq(2, 5, 2))), nrow=5)
   designLatinSqrSys(3)

[Package dae version 3.2.32 Index]