detect.diff {dae}	R Documentation

Computes the detectable difference for an experiment

Description

Computes the delta that is detectable for specified replication, power, alpha.

Usage

detect.diff(rm=5, df.num=1, df.denom=10, sigma=1, alpha=0.05, power=0.8, 
            tol = 0.001, print=FALSE)

Arguments

rm
The number of observations used in computing a mean.

df.num
The degrees of freedom of the numerator of the F for testing the term involving the means.

df.denom
The degrees of freedom of the denominator of the F for testing the term involving the means.

sigma
The population standard deviation.

alpha
The significance level to be used.

power
The minimum power to be achieved.

tol
The maximum difference tolerated between the power required and the power computed in determining the detectable difference.

print
TRUE or FALSE to have or not have a table of power calculation details printed out.

Value

A single numeric value containing the computed detectable difference.

Author(s)

Chris Brien

See Also

power.exp, no.reps in package dae.

Examples

## Compute the detectable difference for a randomized complete block design 
## with four treatments given power is 0.8 and alpha is 0.05. 
rm <- 5
detect.diff(rm = rm, df.num = 3, df.denom = 3 * (rm - 1),sigma = sqrt(20))

[Package dae version 3.2.32 Index]