extab(dae)	R Documentation

Expands the values in table to a vector.

Description

Expands the values in table to a vector according to the index.factors that are considered to index the table, either in standard or yates order. The order of the values in the vector is determined by the order of the values of the index.factors.

Usage

extab(table, index.factors, order="standard")

Arguments

table 	 A numeric vector containing the values to be expanded. Its length must equal the product of the number of used levels for the factors in index.factors and the values in it correspond to all levels combinations of these factors. That is, the values of the index.factors are irrelevant to table.

index.factors 	 A list of factors that index the the table. All the factors must be the same length.

order 	 The order in which the levels combinations of the index.factors are to be considered as numbered in indexing table; standard numbers them as if they are arranged in standard order, that is with the first factor moving slowest and the last factor moving fastest; yates numbers them as if they are arranged in Yates order, that is with the first factor moving fastest and last factor moving slowest.

Value

A vector of length equal to the factors in index.factor whose values are taken from table.

Examples

        AB.effects <- extab(table=AB.tab, index.factors=list(A, B))

[Package dae version 0.4-1 Index]