fitted.errors(dae)	R Documentation

Extract the fitted values for a fitted model

Description

Extracts the fitted values as the sum of the effects for all the fitted terms in the model, stopping at error.term if this is specified.

Usage

fitted.errors(aov.obj, error.term=NULL)

Arguments

aov.obj 	 An aov object or aovlist object created from a call to aov

error.term 	 The error.term down to which effects are extracted for adding to the fitted values

Value

A numeric vector of fitted values

Note

Only terms in the linear model will be included in the fitted values when computing the residuals. Terms in an Error function, but not in the model external to the Error function, will not be included. If you want them included, put them in both.

See Also

resid.errors, tukey.1df.

Examples

        fit  <- fitted.errors(crd.aov)
        rcbd.aov - aov(Y ~ Blocks + Treats + Error(Blocks/Plots))
        fit <- fitted.errors(rcbd.aov, error.term = "Blocks:Plots")

[Package dae version 0.3-1 Index]