get.daeRNGkind {dae}	R Documentation

Gets the value of daeRNGkind for the package dae from the daeEnv environment

Description

A function that gets the character value of daeRNGkind from the daeEnv environment. The value specifies the name of the Random Number Generator to use in dae.

Usage

get.daeRNGkind()

Value

The character value of daeRNGkind.

Author(s)

Chris Brien

See Also

set.daeRNGkind.

Examples

## get daeRNGkind.
get.daeRNGkind()

[Package dae version 3.2.32 Index]