get.daeTolerance {dae}	R Documentation

Gets the value of daeTolerance for the package dae

Description

A function that gets the vector of values such that, in dae functions, values less than it are considered to be zero.

Usage

get.daeTolerance()

Value

The vector of two values for daeTolerance, one named element.tol that is used for elements of matrices and a second named element.eigen that is used for eigenvalues and quantities based on them, such as efficiency factors.

Author(s)

Chris Brien

See Also

set.daeTolerance.

Examples

## get daeTolerance.
get.daeTolerance()

[Package dae version 3.2.32 Index]