interaction.ABC.plot {dae}	R Documentation

Plots an interaction plot for three factors

Description

Plots a function (the mean by default) of the response for the combinations of the three factors specified as the x.factor (plotted on the x axis of each plot), the groups.factor (plotted as separate lines in each plot) and the trace.factor (its levels are plotted in different plots). Interaction plots for more than three factors can be produced by using fac.combine to combine all but two of them into a single factor that is specified as the trace.factor.

Usage

interaction.ABC.plot(response, x.factor, groups.factor, 
       trace.factor,data, fun="mean", title="A:B:C Interaction Plot", 
       xlab, ylab, key.title, lwd=4, columns=2, ggplotFuncs = NULL, ...)

Arguments

response
A numeric vector containing the response variable from which a function (the mean by default) is computed for plotting on the y-axis.

x.factor
The factor to be plotted on the x-axis of each plot. If the levels are numeric values stored as characters, they will be converted to numeric values for plotting. If they are actually numeric codes for nonnumeric categories and you want them plotted on a discrete scale then you should employ nonumeric codings, such as ‘-’ and ‘+’ or ‘N’ and ‘Y’ or something similar.

groups.factor
The factor plotted as separate lines in each plot.

trace.factor
The factor for whose levels there are separate plots.

data
A data.frame containing the three factors and the response.

fun
The function to be computed from the response for each combination of the three factors x.factor, groups.factor and trace.factor. By default, the mean is computed for each combination.

title
Title for plot window. By default it is "A:B:C Interaction Plot".

xlab
Label for the x-axis. By default it is the name of the x.factor.

ylab
Label for the y-axis. By default it is the name of the response.

key.title
Label for the key (legend) to the lines in each plot. By default it is the name of the groups.factor.

lwd
The width of the lines. By default it is 4.

columns
The number of columns for arranging the several plots for the levels of the groups.factor. By default it is 2.

ggplotFuncs
A list, each element of which contains the results of evaluating a ggplot2 function. It is created by calling the list function with a ggplot2 function call for each element. These functions are applied in creating the ggplot object.

...
Other arguments that are passed down to ggplot2 methods.

Value

An object of class "ggplot", which can be plotted using print.

Author(s)

Chris Brien

See Also

fac.combine in package dae, interaction.plot.

Examples

## Not run: 
## plot for Example 14.1 from Mead, R. (1990). The Design of Experiments: 
## Statistical Principles for Practical Application. Cambridge, 
## Cambridge University Press.  
## use ?SPLGrass.dat for details
data(SPLGrass.dat)
interaction.ABC.plot(Main.Grass, x.factor=Period,
                     groups.factor=Spring, trace.factor=Summer,
                     data=SPLGrass.dat,
                     title="Effect of Period, Spring and Summer on Main Grass")

## plot for generated data
## use ?ABC.Interact.dat for data set details
data(ABC.Interact.dat)
## Add standard errors for plotting 
## - here data contains a single value for each combintion of A, B and C
## - need to supply name for data twice 
ABC.Interact.dat$se <- rep(c(0.5,1), each=4)
interaction.ABC.plot(MOE, A, B, C, data=ABC.Interact.dat,
                     ggplotFunc=list(geom_errorbar(data=ABC.Interact.dat, 
                                                   aes(ymax=MOE+se, ymin=MOE-se), 
                                                   width=0.2)))

## End(Not run)

[Package dae version 3.2.32 Index]