mat.I {dae}	R Documentation

Forms a unit matrix

Description

Form the unit or identity matrix of order order.

Usage

mat.I(order)

Arguments

order
The order of the matrix to be formed.

Value

A square matrix whose diagonal elements are one and its off-diagonal are zero.

Author(s)

Chris Brien

See Also

mat.J, mat.ar1

Examples

    col.I <- mat.I(order=4)

[Package dae version 3.2.32 Index]