mat.J {dae}	R Documentation

Forms a square matrix of ones

Description

Form the square matrix of ones of order order.

Usage

mat.J(order)

Arguments

order
The order of the matrix to be formed.

Value

A square matrix all of whose elements are one.

Author(s)

Chris Brien

See Also

mat.I, mat.ar1

Examples

    col.J <- mat.J(order=4)

[Package dae version 3.2.32 Index]