mat.ar1 {dae}	R Documentation

Forms an ar1 correlation matrix

Description

Form the correlation matrix of order order whose correlations follow the ar1 pattern. The matrix is banded and has diagonal elements equal to one and the off-diagonal element in the ith row and jth column equal to rho^(k) where k = abs(i- j).

Usage

mat.ar1(rho, order)

Arguments

rho
The correlation on the first off-diagonal.

order
The order of the matrix to be formed.

Value

A banded correlation matrix whose elements follow an ar1 pattern.

Author(s)

Chris Brien

See Also

mat.I, mat.J, mat.cor, mat.corg, mat.exp, mat.gau, mat.banded, mat.ar2, mat.ar3, mat.sar2, mat.ma1, mat.ma2, mat.arma

Examples

    corr <- mat.ar1(rho=0.4, order=4)

[Package dae version 3.2.32 Index]