mat.cor {dae}	R Documentation

Forms a correlation matrix in which all correlations have the same value.

Description

Form the correlation matrix of order order in which all correlations have the same value.

Usage

mat.cor(rho, order)

Arguments

rho
A numeric containing the single correlation value.

order
The order of the correlation matrix to be formed.

Value

A correlation matrix.

Author(s)

Chris Brien

See Also

mat.I, mat.J, mat.corg, mat.banded, mat.exp, mat.gau, mat.ar1, mat.ar2, mat.sar2, mat.ma1, mat.ma2, mat.arma

Examples

    corr <- mat.cor(rho = 0.4, order = 3)

[Package dae version 3.2.32 Index]