mat.corg {dae}	R Documentation

Forms a general correlation matrix

Description

Form the correlation matrix of order order for which all correlations potentially differ.

Usage

mat.corg(rhos, order, byrow = FALSE)

Arguments

rhos
A numeric containing the p(p -1)/2 correlation values ordered either by columns (if byrow is FALSE) or by rows (if byrow is TRUE).

order
The order of the correlation matrix to be formed.

byrow
A logical. If FALSE the lower-traingle of the matrix is filled by columns, otherwise the the ower triangle is filled by rows.

Value

A correlation matrix.

Author(s)

Chris Brien

See Also

mat.I, mat.J, mat.cor, mat.banded, mat.exp, mat.gau, mat.ar1, mat.ar2, mat.sar2, mat.ma1, mat.ma2, mat.arma

Examples

    corr <- mat.corg(rhos = c(0.4, 0.2, 0.1), order = 3)

[Package dae version 3.2.32 Index]