mat.dirprod {dae}	R Documentation

Forms the direct product of two matrices

Description

Form the direct product of the m x n matrix A and the p x q matrix B. It is also called the Kroneker product and the right direct product. It is defined to be the result of replacing each element of A, a_{ij}, with a_{ij}\bold{B}. The result matrix is mp \times nq.

The method employed uses the rep function to form two mp \times nq matrices: (i) the direct product of A and J, and (ii) the direct product of J and B, where each J is a matrix of ones whose dimensions are those required to produce an mp \times nq matrix. Then the elementwise product of these two matrices is taken to yield the result.

Usage

mat.dirprod(A, B)

Arguments

A
The left-hand matrix in the product.

B
The right-hand matrix in the product.

Value

An mp \times nq matrix.

Author(s)

Chris Brien

See Also

matmult, mat.dirprod

Examples

    col.I <- mat.I(order=4)
    row.I <- mat.I(order=28)
    V <- mat.dirprod(col.I, row.I)

[Package dae version 3.2.32 Index]