mat.gau {dae}	R Documentation

Forms an exponential correlation matrix

Description

Form the correlation matrix of order equal to the length of coordinates. The matrix has diagonal elements equal to one and the off-diagonal element in the ith row and jth column equal to rho^(k) where k = (coordinate[i]- coordinate[j])^2.

Usage

mat.gau(rho, coordinates)

Arguments

rho
The correlation for points a distance of one apart.

coordinates
The coordinates of points whose correlation matrix is to be formed.

Value

A correlation matrix whose elements depend on the power of the absolute distance apart.

Author(s)

Chris Brien

See Also

mat.I, mat.J, mat.cor, mat.corg, mat.banded, mat.ar1, mat.ar2, mat.ar3, mat.sar2, mat.ma1, mat.ma2, mat.arma, mat.exp

Examples

    corr <- mat.gau(coordinates=c(3:6, 9:12, 15:18), rho=0.1)

[Package dae version 3.2.32 Index]