mat.ma2 {dae}	R Documentation

Forms an ma2 correlation matrix

Description

Form the correlation matrix of order order whose correlations follow the ma2 pattern. The resulting matrix is banded.

Usage

mat.ma2(MAparameters, order)

Arguments

MAparameters
A numeric containing the two moving average parameter values of the process, being the weights given to the lag 1 and lag 2 random pertubations.

order
The order of the matrix to be formed.

Details

The correlations in the correlation matrix, corr say, are calculated from the moving average parameters, MAparameters. The values in

the diagonal (k = 1) of corr are one;

the first subdiagonal band (k = 2) of corr are equal to
-MAparameters[1]*(1 - MAparameters[2]) / div;

the second subdiagonal bande (k = 3) of corr are equal to -MAparameters[2] / div;

in subsequent disgonal bands, (k = 4:order), of corr are zero,

where div = 1 + MMAparameters[1]*MAparameter[1] + MAparameters[2]*MAparameters[2].

Value

A banded correlation matrix whose elements follow an ma2 pattern.

Author(s)

Chris Brien

See Also

mat.I, mat.J, mat.cor, mat.corg, mat.exp, mat.gau, mat.banded, mat.ar1, mat.ar3, mat.sar2, mat.ma1, mat.arma

Examples

    corr <- mat.ma2(MAparameters = c(0.4, -0.2), order = 4)

[Package dae version 3.2.32 Index]