mat.sar {dae}	R Documentation

Forms an sar correlation matrix

Description

Form the correlation matrix of order order whose correlations follow the sar pattern. The resulting matrix is banded.

Usage

mat.sar(SARparameter, order)

Arguments

SARparameter
A numeric containing the single value of the parameter from which the correlations are calculated.

order
The order of the matrix to be formed.

Details

The values of the correlations in the correlation matrix, corr say, are calculated from the SARparameter, gamma as follows. The values in

the diagonal of corr (k = 1) are one;

the first subdiagonal band (k = 2) of corr are equal to gamma/(1 + (gamma * gamma / 4));

the subsequent subdiagonal bands, (k = 3:order), of corr are equal to
gamma * corr[k-1] - (gamma * gamma/4) * corr[k-2].

Value

A banded correlation matrix whose elements follow an sar pattern.

Author(s)

Chris Brien

See Also

mat.I, mat.J, mat.cor, mat.corg, mat.banded, mat.exp, mat.gau, mat.ar1, mat.ar2, mat.ar3, mat.sar2, mat.ma1, mat.ma2, mat.arma

Examples

    corr <- mat.sar(SARparameter = -0.4, order = 4)

[Package dae version 3.2.32 Index]