p2canon.object {dae}	R Documentation

Description of a p2canon object

Description

An object of class p2canon that contains information derived from two formulae using projs.2canon.

Value

A list of class p2canon. It has two components: decomp and aliasing. The decomp component iscomposed as follows:

It has a component for each component of Q1.

Each of the components for Q1 is a list; each of these lists has one component for each of Q2 and a component Pres.

Each of the Q2 components is a list of three components: pairwise, adjusted and Qproj. These components are based on an eigenalysis of the relationship between the projectors for the parent Q1 and Q2 components.

Each pairwise component is based on the nonzero canonical efficiency factors for the joint decomposition of the two parent projectors (see proj2.eigen).

An adjusted component is based on the nonzero canonical efficiency factors for the joint decomposition of the Q1 component and the Q2 component, the latter adjusted for all Q2 projectors that have occured previously in the list.

The Qproj component is the adjusted projector for the parent Q2 component.

The pairwise and adjusted components have the following components: efficiencies, aefficiency, mefficiency, sefficiency, eefficiency, xefficiency, order and dforthog – for details see efficiency.criteria.

The aliasing component is a data.frame decribing the aliasing between terms corresponding to two Q2 projectors when estimated in subspaces corresponding to a Q1 projector.

Author(s)

Chris Brien

See Also

projs.2canon, designAnatomy, pcanon.object.

[Package dae version 3.2.32 Index]