pcanon.object {dae}	R Documentation

Description of a pcanon object

Description

An object of class pcanon that contains information derived from several formulae using designAnatomy.

Value

A list of class pcanon that has four components: (i) Q, (ii) terms, (iii) sources, (iv) marginality, and (v) aliasing. Each component is a list with as many components as there are formulae in the formulae list supplied to designAnatomy.

The Q list is made up of the following components:

The first component is the joint decomposition of two structures derived from the first two formulae, being the p2canon.object produced by projs.2canon.

Then there is a component for each further formulae; it contains the p2canon.object obtained by applying projs.2canon to the structure for a formula and the already established joint decomposition of the structures for the previous formulae in the formulae.

The last component contains the the list of the projectors that give the combined canonical decomposition derived from all of the formulae.

The terms, sources, marginalty and aliasing lists have a component for each formula in the formulae argument to designAnatomy, Each component of the terms and sources lists has a character vector containing the terms or sources derived from its formula. For the marginality component, each component is the marginality matrix for the terms derived from its formula. For the aliasing component, each component is the aliasing data.frame for the source derived from its formula. The components of these four lists are produced by pstructure.formula and are copied from the pstructure.object for the formula. The names of the components of these four lists will be the names of the components in the formulae list.

The object has the attribute labels, which is set to "terms" or "sources" according to which of these were used to label the projectors when the object was created.

Author(s)

Chris Brien

See Also

designAnatomy, p2canon.object.

[Package dae version 3.2.32 Index]