power.exp(dae)	R Documentation

Computes the power for an experiment.

Description

Computes the power for an experiment.

Usage

power.exp(rm=5., df.num=1., df.denom=10., delta=1., sigma=1.,
                 alpha=0.05, print=FALSE)

Arguments

rm 	 The number of observations used in computing a mean.

df.num 	 The degrees of freedom of the numerator of the F for testing the term involving the means.

df.denom 	 The degrees of freedom of the denominator of the F for testing the term involving the means.

delta 	 The true difference between a pair of means.

sigma 	 The population standard deviation.

alpha 	 The significance level to be used.

print 	 TRUE or FALSE to have or not have a table of power calculation details printed out.

Value

A scalar containing the computed power.

See Also

no.reps

Examples

        rm <- 5
        power.exp(rm = rm, df.num = 3, df.denom = 3 * (rm - 1), delta = 5,
            sigma = sqrt(20),power = 0.8, print = TRUE)

[Package dae version 0.3-1 Index]