print.aliasing {dae}	R Documentation

Print an aliasing data.frame

Description

Prints an aliasing data.frame.

Usage

## S3 method for class 'aliasing'
print(x, which.criteria = c("aefficiency","eefficiency","order"),  ...)

Arguments

x
The data.frame that is also of class aliasing and is to be printed.

which.criteria
A character vector nominating the efficiency criteria to be included in the summary of aliasing between terms. It can be none, all or some combination of aefficiency, mefficiency, sefficiency, eefficiency, xefficiency, order and dforthog – for details see efficiency.criteria. If none, no criteria are printed.

...
Further arguments passed to the print method for data.frame.

Author(s)

Chris Brien

See Also

print, print.default, show.

Examples

## Generate a data.frame with 3 factors length 12
pseudo.lay <- data.frame(pl = factor(1:12),
                         ab = factor(rep(1:4, times=3)),
                         a = factor(rep(1:2, times=6)))

## create a pstructure object
trt.struct <- pstructure(~ ab+a, data = pseudo.lay)

## print the object either using the Method function, the generic function or show
print.aliasing(trt.struct$aliasing)
print(trt.struct$aliasing, which.criteria = "none")
trt.struct$aliasing

[Package dae version 3.2.32 Index]