print.projector {dae}	R Documentation

Print projectors

Description

Print an object of class "projector", displaying the matrix and its degrees of freedom (rank).

Usage

## S3 method for class 'projector'
print(x, ...)

Arguments

x
The object of class "projector" to be printed.

...
Further arguments passed to or from other methods.

Author(s)

Chris Brien

See Also

print, print.default, show.

projector for further information about this class.

Examples

## set up a 2 x 2 mean operator that takes the mean of a vector of 2 values
m <- matrix(rep(0.5,4), nrow=2)

## create an object of class projector
proj.m <- projector(m)

## print the object either using the Method function, the generic function or show
print.projector(proj.m)
print(proj.m)
proj.m

[Package dae version 3.2.32 Index]