print.pstructure {dae}	R Documentation

Prints a pstructure.object

Description

Prints a pstructure.object, which is of class pstructure. The df, terms and sources are coerced into a data.frame and printed; the marginality matrix is printed separately.

Usage

## S3 method for class 'pstructure'
print(x, which = "all", ...)

Arguments

x
The pstructure.object, which is of class pstructure and is to be printed.

which
A character vector nominating the components of the pstructure.object to print. Must be all or some combination of projectors, marginality, and aliasing.

...
Further arguments passed to print.aliasing.

Author(s)

Chris Brien

See Also

print, print.default, show.

Examples

## Generate a data.frame with 4 factors, each with three levels, in standard order
ABCD.lay <- fac.gen(list(A = 3, B = 3, C = 3, D = 3))

## create a pstructure object based on the formula ((A*B)/C)*D
ABCD.struct <- pstructure.formula(~ ((A*B)/C)*D, data =ABCD.lay)

## print the object either using the Method function, the generic function or show
print.pstructure(ABCD.struct)
print(ABCD.struct)
ABCD.struct

[Package dae version 3.2.32 Index]