pstructure.object {dae}	R Documentation

Description of a pstructure object

Description

An object of class pstructure that contains information derived from a formula using pstructure.formula. It also inherits from class list.

Value

A list of class pstructure with the following components:

Q: a list with a component of class projector, being the orthogonalized projectors for each non-aliased term/source in the formula; if grandMean is TRUE in the call to pstructure.formula then it also includes the projector for it;

terms: a character vector with the non-aliased term names; if grandMean is TRUE in the call to pstructure.formula then the first term will be "Mean";

sources: a character vector with the non-aliased source names;

marginality: a matrix of zeroes and ones with the same number of rows and columns as number of non-aliased terms, excluding the term for the grand mean even when grandMean is TRUE; the row names and column names are the elements terms, excluding "Mean";

the entry in the ith row and jth column will be one if the ith term is marginal to the jth term i.e. the column space of the ith term is a subspace of that for the jth term and so the source for the jth term will have been made orthogonal to that for the ith term; otherwise, the entries are zero.

aliasing: a data.frame containing the information about the (partial) aliasing between the sources in the formula. The columns are:

Source: the source names, or associated term name, for those that are (partially) aliased with previous sources;

df: the remaining degrees of freedom for the source;

Alias: the source with which the current entry is (partially) aliased;

efficiency criteria: a set of columns for the complete set of criteria calculated by efficiency.criteria; the criteria reflect the amount of information that is aliased with previous sources and a line is included in the component that reports the informaton remaining after adjustment for previous sources.

The information provided depends on the setting of orthogonalize. All the information is provided for the "hybrid" option. For the option "differencing", no efficiency criteria are included and either the terms/sources of the Alias are set to "unknown" and the df are set to NA when these are unknown. For the option "eigenmethods", the previous terms/sources cannot be identified and so all values of Alias are set to NA. If there is no (partial) aliasing then the component is set to NULL.

The object has the attribute labels, which is set to "terms" or "sources" according to which of these label the projectors.

Author(s)

Chris Brien

See Also

pstructure.formula and, for further information about the projector classs, projector.

[Package dae version 3.2.32 Index]