qqnorm.yeffects(dae)	R Documentation

Half or full normal plot of Yates effects.

Description

Produces a half or full normal plot of the Yates effects from a 2^k factorial experiment.

Usage

qqnorm.yeffects(aov.obj, error.term="Within", data=NULL, pch=16, full=FALSE, ...)

Arguments

aov.obj 	 An aov object or aovlist object created from a call to aov.

error.term 	 The error.term from which the Yates effects are estimated. Only required when Error used in call to aov.

data 	 A data frame in which the variables specified in the aov.obj will be found. If missing, the variables are searched for in the standard way.

pch 	 the number of a plotting symbol to be drawn when plotting points (use help(points) for details).

full 	 whether a full or half normal plot is to be produced. The default is for a half-normal plot; full=TRUE produces a full normal plot.

... 	 Further graphical parameters may be specified (use help(par) for possibilities.

Details

A half or full normal plot of the Yates effects is produced. You will be able to interactively select effects to be labelled (click on the side of the point where you want the label placed). Right click on the graph and select Stop when you have finished labelling effects. A regression line fitted to the unselected effects and constrained to go through the origin is plotted. Also, a list of the labelled effects, if any, are printed to standard ouptut.

Value

returns, invisibly, a list with components x and y, giving coordinates of the plotted points.

See Also

yates.effects, qqnorm.

Examples

  Fac4Proc.aov <- aov(Conv ~ Catal * Temp * Press * Conc + Error(Runs),
                                                            Fac4Proc.dat)
  qqnorm.yeffects(Fac4Proc.aov, error.term="Runs", data=Fac4Proc.dat)

[Package dae version 0.3-4 Index]