qqyeffects {dae}	R Documentation

Half or full normal plot of Yates effects

Description

Produces a half or full normal plot of the Yates effects from a 2^k factorial experiment.

Usage

qqyeffects(aov.obj, error.term="Within", data=NULL, pch=16, 
           full=FALSE, ...)

Arguments

aov.obj
An aov object or aovlistobject created from a call to aov.

error.term
The term from the Error function from which the Yates effects are estimated. Only required when Error used in call to aov.

data
A data.frame in which the variables specified in the aov.obj will be found. If missing, the variables are searched for in the standard way.

pch
The number of a plotting symbol to be drawn when plotting points (use help(points) for details).

full
whether a full or half normal plot is to be produced. The default is for a half-normal plot; full=TRUE produces a full normal plot.

...
Further graphical parameters may be specified (use help(par) for possibilities.

Details

A half or full normal plot of the Yates effects is produced. You will be able to interactively select effects to be labelled (click reasonably close to the point and on the side where you want the label placed). Right click on the graph and select Stop when you have finished labelling effects. A regression line fitted to the unselected effects and constrained to go through the origin is plotted. Also, a list of the labelled effects, if any, are printed to standard ouptut.

Value

Returns, invisibly, a list with components x and y, giving coordinates of the plotted points.

Author(s)

Chris Brien

See Also

yates.effects in package dae, qqnorm.

Examples

## analysis of 2^4 factorial experiment from Table 10.6 of Box, Hunter and 
## Hunter (1978) Statistics for Experimenters. New York, Wiley.
## use ?Fac4Proc.dat for data set details
data(Fac4Proc.dat)
Fac4Proc.aov <- aov(Conv ~ Catal * Temp * Press * Conc + Error(Runs),
                                                            Fac4Proc.dat)
qqyeffects(Fac4Proc.aov, error.term="Runs", data=Fac4Proc.dat)

[Package dae version 3.2.32 Index]