resid.errors {dae}	R Documentation

Extract the residuals for a fitted model

Description

An alias for the generic function residuals. When it is available, the method residuals.aovlist extracts residuals, which is provided in the package dae to cover aovlist objects.

Usage

resid.errors(...)

Arguments

...
Arguments passed to residuals.aovlist.

Value

A numeric vector containing the residuals.

Note

See residuals.aovlist for specific information about the residuals when an Error function is used in the call to the aov function.

Author(s)

Chris Brien

See Also

fitted.errors, residuals.aovlist, tukey.1df in package dae.

Examples

## set up data frame for randomized complete block design in Table 4.4 from 
## Box, Hunter and Hunter (2005) Statistics for Experimenters. 2nd edn 
## New York, Wiley.
RCBDPen.dat <- fac.gen(list(Blend=5, Flask=4))
RCBDPen.dat$Treat <- factor(rep(c("A","B","C","D"), times=5))
RCBDPen.dat$Yield <- c(89,88,97,94,84,77,92,79,81,87,87,
                       85,87,92,89,84,79,81,80,88)

## perform the analysis of variance
RCBDPen.aov <- aov(Yield ~ Blend + Treat + Error(Blend/Flask), RCBDPen.dat)
summary(RCBDPen.aov)

## two equivalent ways of extracting the residuals
res  <- residuals.aovlist(RCBDPen.aov)
res <- residuals(RCBDPen.aov, error.term = "Blend:Flask")
res <- resid.errors(RCBDPen.aov)

[Package dae version 3.2.32 Index]