residuals.aovlist {dae}	R Documentation

Extract the residuals from an aovlist object

Description

Extracts the residuals from error.term or, if error.term is not specified, the last error.term in the analysis. It is a method for the generic function residuals.

Usage

## S3 method for class 'aovlist'
residuals(object, error.term=NULL, ...)

Arguments

object
An aovlist object created from a call to aov.

error.term
The term from the Error function for which the residuals are to be extracted. If error.term is NULL the residuals are extracted from the last Error term.

...
Further arguments passed to or from other methods.

Value

A numeric vector containing the residuals.

Author(s)

Chris Brien

See Also

fitted.errors, resid.errors, tukey.1df in package dae.

Examples

## set up data frame for randomized complete block design in Table 4.4 from 
## Box, Hunter and Hunter (2005) Statistics for Experimenters. 2nd edn 
## New York, Wiley.
RCBDPen.dat <- fac.gen(list(Blend=5, Flask=4))
RCBDPen.dat$Treat <- factor(rep(c("A","B","C","D"), times=5))
RCBDPen.dat$Yield <- c(89,88,97,94,84,77,92,79,81,87,87,
                       85,87,92,89,84,79,81,80,88)

## perform the analysis of variance
RCBDPen.aov <- aov(Yield ~ Blend + Treat + Error(Blend/Flask), RCBDPen.dat)
summary(RCBDPen.aov)

## two equivalent ways of extracting the residuals
res  <- residuals.aovlist(RCBDPen.aov)
res <- residuals(RCBDPen.aov, error.term = "Blend:Flask")

[Package dae version 3.2.32 Index]