set.daeRNGkind {dae}	R Documentation

Sets the values of daeRNGkind for the package dae in the daeEnv environment

Description

A function that sets the character value daeRNGkind that specifies the kind of Random Number generator to use in dae. The value is stored in a character named daeRNGkind in the daeEnv environment. It is initially set to "Mersenne-Twister" and can be changed using get.daeRNGkind. For details of the different Random Number Generators available in R, see the R help for RNGkind.

Usage

set.daeRNGkind(kind = "Mersenne-Twister")

Arguments

kind
A character to which daeRNGkind is to be set.

Value

The value of daeRNGkind is returned invisibly.

Author(s)

Chris Brien

See Also

get.daeRNGkind.

Examples

## set daeRNGkind to L'Ecuyer-CMRG.
set.daeRNGkind("L'Ecuyer-CMRG")

[Package dae version 3.2.32 Index]