set.daeTolerance {dae}	R Documentation

Sets the values of daeTolerance for the package dae

Description

A function that sets the values such that, in dae functions, values less than it are considered to be zero. The values are stored in a vector named daeTolerance in the daeEnv environment. The vector is of length two and, initially, both values are set to .Machine$double.eps ^ 0.5 (about 1.5E-08). One value is named element.tol and is used for elements of matrices; the second is named element.eigen and is used for eigenvalues and quantities based on them, such as efficiency factors.

Usage

set.daeTolerance(element.tol=NULL, eigen.tol=NULL)

Arguments

element.tol
The value to to which the first element of the daeTolerance vector is to be set. If more than one value is supplied, only the first value is used.

eigen.tol
The value to to which the second element of the daeTolerance vector is to be set. If more than one value is supplied, only the first value is used.

Value

The vector daeTolerance is returned invisibly.

Author(s)

Chris Brien

See Also

get.daeTolerance.

Examples

## set daeTolerance.
set.daeTolerance(1E-04, 1E-08)

[Package dae version 3.2.32 Index]