summary.p2canon {dae}	R Documentation

Summarize a canonical analysis of the relationships between two sets of projectors

Description

Produces a summary of the efficiency criteria computed from the canonical efficiency factors for the joint decomposition of two sets of projectors (Brien and Bailey, 2009) obtained using projs.2canon. It takes the form of a decomposition or skeleton ANOVA table.

Usage

## S3 method for class 'p2canon'
summary(object, which.criteria = c("aefficiency", "eefficiency", "order"), ...)

Arguments

object
A list of class p2canon produced by projs.2canon.

which.criteria
A character vector nominating the efficiency criteria to be included in the summary. It can be none, all or some combination of aefficiency, mefficiency, sefficiency, eefficiency, xefficiency, order and dforthog – for details see efficiency.criteria.

...
further arguments affecting the summary produced.

Value

An object of classes summary.p2canon and data.frame, whose rows correspond to the pairs of projectors, one from the Q1 argument and the other from the Q2 argument from projs.2canon; only pairs with non-zero efficiency factors are included. In addition, a line is included for each nonzero Residual Q1 projector.

Author(s)

Chris Brien

References

Brien, C. J. and R. A. Bailey (2009). Decomposition tables for multitiered experiments. I. A chain of randomizations. The Annals of Statistics, 36, 4184 - 4213.

See Also

projs.2canon, proj2.efficiency, efficiency.criteria, proj2.combine, proj2.eigen, pstructure,
print.summary.p2canonin package dae, eigen.

projector for further information about this class.

Examples

## PBIBD(2) from p. 379 of Cochran and Cox (1957) Experimental Designs. 
## 2nd edn Wiley, New York
PBIBD2.unit <- list(Block = 6, Unit = 4)
PBIBD2.nest <- list(Unit = "Block")
trt <- factor(c(1,4,2,5, 2,5,3,6, 3,6,1,4, 4,1,5,2, 5,2,6,3, 6,3,4,1))
PBIBD2.lay <- designRandomize(allocated = trt, 
                              recipient = PBIBD2.unit, 
                              nested.recipients = PBIBD2.nest)

##obtain projectors using pstructure
unit.struct <- pstructure(~ Block/Unit, data = PBIBD2.lay)
trt.struct <- pstructure(~ trt, data = PBIBD2.lay)

##obtain combined decomposition and summarize
unit.trt.p2canon <- projs.2canon(unit.struct$Q, trt.struct$Q)
summary(unit.trt.p2canon)

[Package dae version 3.2.32 Index]