summary.pcanon {dae}	R Documentation

Summarizes the anatomy of a design, being the decomposition of the sample space based on its canonical analysis, as produced by designAnatomy

Description

Gives the anatomy of a design in a table; it summarizes the joint decomposition of two or more sets of projectors (Brien and Bailey, 2009) obtained using designAnatomy. It includes the efficiency criteria computed from the canonical efficiency factors for the joint decomposition. The labels in the table may be Terms or Sources. The terms are those that would be included in a mixed model for an experiment based on the design. The sources are the orthogonal subspaces, derived from the terms, that make up the decomposition and the degrees of freedom and efficiency factors relate to these subspaces. The table displays how the information for the different sources allowed for in the design are related. For more information about the notation used for sources see the labels argument of designAnatomy.

It is possible to supply an object that is a pcanon.object produced in versions prior to 3.0-0 using projs.canon.

Usage

## S3 method for class 'pcanon'
summary(object, labels.swap = FALSE, 
        which.criteria = c("aefficiency", "eefficiency", "order"), ...)

Arguments

object
A pcanon.object.

labels.swap
A logical indicating whether to swap between "sources" and ‘terms’ in the output. The default is established by the labels argument of designAnatomy and projs.canon.

which.criteria
A character vector nominating the efficiency criteria to be included in the summary. It can be none, all or some combination of aefficiency, mefficiency, sefficiency, eefficiency, xefficiency, order and dforthog – for details see efficiency.criteria. If there is only one formula, this argument is ignored.

...
further arguments affecting the summary produced.

Value

An object of class summary.pcanon that is a list with the two components decomp and aliasing.

The component decomp is a data.frame whose rows correspond to subspaces in the decomposition for a design. It has the following attributes: (i) title that is the title for printing with the decomposition table; (ii) ntiers that is equal to the number of tiers; (iii) orthogonal that is TRUE if the design is orthogonal; (iv) labels that is either "terms" or "sources" depending on the labels that have resulted from the setting of label.swap.

The component aliasing is a data.frame that is also of class aliasing. It contains information about the aliasing between terms that are derived from the same formula and has the attribute title that is the title to be printed with the aliasing table.

However, if the object supplied is a pcanon.object produced with versions prior to 3.0-0 using projs.canon, the value is a data.frame, instead of a list, that has the same attributes as the decomp component of the summary.pcanon object now produced, except that labels is always set to "terms".

Author(s)

Chris Brien

References

Brien, C. J. and R. A. Bailey (2009). Decomposition tables for multitiered experiments. I. A chain of randomizations. The Annals of Statistics, 36, 4184 - 4213.

See Also

designAnatomy, designAnatomy, , pstructure, efficiency.criteria, proj2.combine,
proj2.efficiency, proj2.eigen, print.summary.pcanonin package dae, eigen.

projector for further information about this class.

Examples

## PBIBD(2) from p. 379 of Cochran and Cox (1957) Experimental Designs. 
## 2nd edn Wiley, New York
PBIBD2.unit <- list(Block = 6, Unit = 4)
PBIBD2.nest <- list(Unit = "Block")
trt <- factor(c(1,4,2,5, 2,5,3,6, 3,6,1,4, 4,1,5,2, 5,2,6,3, 6,3,4,1))
PBIBD2.lay <- designRandomize(allocated = trt, 
                              recipient = PBIBD2.unit, 
                              nested.recipients = PBIBD2.nest)

##obtain combined decomposition and summarize
unit.trt.canon <- designAnatomy(list(unit=~ Block/Unit, trt=~ trt), 
                                data = PBIBD2.lay)
summary(unit.trt.canon, which = c("aeff","eeff","order"))
summary(unit.trt.canon, which = c("aeff","eeff","order"), labels.swap = TRUE)

[Package dae version 3.2.32 Index]