yates.effects {dae}	R Documentation

Extract Yates effects

Description

Extracts Yates effects from an aov object or aovlist object.

Usage

yates.effects(aov.obj, error.term="Within", data=NULL)

Arguments

aov.obj
An aov object or aovlist object created from a call to aov.

error.term
The term from the Error function from which the Yates effects are estimated. Only required when Error used in call to aov.

data
A data.frame in which the variables specified in the aov.obj will be found. If missing, the variables are searched for in the standard way.

Details

Yates effects are specific to 2^k experiments, where Yates effects are conventionally defined as the difference between the upper and lower levels of a factor. We follow the convention used in Box, Hunter and Hunter (1978) for scaling of higher order interactions: all the Yates effects are on the same scale, and represent the average difference due to the interaction between two different levels. Effects are estimated only from the error term supplied to the error.term argument.

Value

A vector of the Yates effects.

Author(s)

Chris Brien

See Also

qqyeffects in package dae, aov.

Examples

## analysis of 2^4 factorial experiment from Table 10.6 of Box, Hunter and 
## Hunter (1978) Statistics for Experimenters. New York, Wiley.
## use ?Fac4Proc.dat for data set details
data(Fac4Proc.dat)
Fac4Proc.aov <- aov(Conv ~ Catal * Temp * Press * Conc + Error(Runs),
                                                            Fac4Proc.dat)
round(yates.effects(Fac4Proc.aov, error.term="Runs", data=Fac4Proc.dat), 2)

[Package dae version 3.2.32 Index]