fac.recode(dae)	R Documentation

Recodes the values of a factor using each value in a supplied vector to replace the corresponding level of the factor.

Description

Recodes factor values using values in a vector.

Usage

fac.recode(factor, newlevels)

Arguments

factor 	 The factor to be recoded.

newlevels 	 A vector of length levels(factor) containing values to use in the recoding.

Value

A factor.

See Also

factor

Examples

  a <- factor(rep(1:4, 4))
        b <- fac.recode(a, c(1,2,2,1))

[Package dae version 0.4-2 Index]