no.reps(dae)	R Documentation

Computes the number of replicates for an experiment.

Description

Computes the number of pure replicates required in an experiment to achieve a specified power.

Usage

no.reps(multiple=1., df.num=1.,
               df.denom=expression((df.num + 1.) * (r - 1.)), delta=1.,
               sigma=1., alpha=0.05, power=0.8, tol=0.025, print=FALSE)

Arguments

multiple 	 The multiplier, m, which when multiplied by the number of pure replicates of a treatment, r, gives the number of observations rm used in computing means for some, not necessarily proper, subset of the treatment factors; m is the replication arising from other treatment factors. However, for single treatment factor experiments the subset can only be the treatment factor and m = 1.

df.num 	 The degrees of freedom of the numerator of the F for testing the term involving the treatment factor subset.

df.denom 	 The degrees of freedom of the denominator of the F for testing the term involving the treatment factor subset.

delta 	 The true difference between a pair of means for some, not necessarily proper, subset of the treatment factors.

sigma 	 The population standard deviation.

alpha 	 The significance level to be used.

power 	 The minimum power to be achieved.

tol 	 The maximum difference tolerated between the power required and the power computed in determining the number of replicates.

print 	 TRUE or FALSE to have or not have a table of power calculation details printed out.

Value

A scalar containing the computed number of pure replicates

See Also

power.exp

Examples

        no.reps(multiple = 1, df.num = 3,
         df.denom = expression(df.num * (r - 1)), delta = 5,
               sigma = sqrt(20), print = TRUE)

[Package dae version 0.3-1 Index]