strength(dae)	R Documentation

Generate paper strength values

Description

Generates paper strength values for an experiment with different temperatures.

Usage

strength(nodays, noruns, temperature, ident)

Arguments

nodays 	 The number of days over which the experiment is to be run.

noruns 	 The number of runs to be performed on each day of the experiment.

temperature 	 A factor that encapsulates the layout by giving the temperature to be investigated for each run on each day. These must be ordered so that the temperatures for the first day are given in the order in which they are to be investigated on that day. These must be followed by the noruns temperatures for the second day and so on. Consequently, the factor temperature will have nodays*noruns values.

ident 	 The digits of your student identity number. That is, leave out any letters.

Value

A data.frame object containing the factors day, run and temperature and the generated strengths.

Examples

        library(dae)
        temperature <- factor( ... )
        exp.strength <- strength(nodays = 4, noruns = 5,
                           temperature = temperature, ident = 0123456)

  Here temperature is a factor with 4*5 = 20 values whose
  first 5 values specify the temperatures to be applied in
  the 5 runs on the first day, values 6 to 10 specify the
  temperatures for the 5 runs on day 2, and so on.

        exp.strength <- strength(nodays = 4, noruns = 5,
                           temperature = Design$temperature, ident = 0123456)

  In the second example, Design is the name of a data.frame object
  that includes the factor temperature.

[Package dae version 0.1-1 Index]