tukey.1df(dae)	R Documentation

Performs Tukey's one-degree-of-freedom-test-for-nonadditivity

Description

Performs Tukey's one-degree-of-freedom-test-for-nonadditivity on a set of residuals from an analysis of variance

Usage

tukey.1df(aov.obj, data, error.term="Within")

Arguments

aov.obj 	 An aov object or aovlist object created from a call to aov.

error.term 	 The error.term whose residuals are to be tested for nonadditivity.

data 	 A data.frame containing the original response variable and factors used in the call to aov.

Value

A list containing Tukey.SS, Tukey.F, Tukey.p, Devn.SSq being the SSq for the 1df test, F value for test and the p-value for the test for the 1df test, F value for test and the p-value for the test.

Note

Only terms in the linear model will be included in the fitted values and so will not be excluded from the residuals. In partiicular, terms in an Error function, but not in the model external to the Error function, will not be included. To include them, put them in both.

See Also

resid.errors, fitted.errors.

Examples

        rcbd.aov - aov(Y ~ Blocks + Treats + Error(Blocks/Plots))
        tukey.1df(rcbd.aov, rcbd, error.term = "Blocks:Plots")

[Package dae version 0.3-1 Index]